Add option to clobber if an image exists with the same name.

+++ Thumbnail.py
@@ -17,16 +17,25 @@ class Thumbnail(ndb.Model):
         return images.resize(pre, 90, 90)
 
     @staticmethod
-    def add_image(filename, image, size):
+    def add_image(filename, image, size, clobber=False):
         # Write the file to cloud storage
         bucket_name = app_identity.get_default_gcs_bucket_name()
         filename = "/%s/%s.png" % (bucket_name, filename)
-        gcs_file = gcs.open(filename, "w", content_type="image/png")
-        gcs_file.write(image)
-        gcs_file.close()
+
+        try:
+            stat = gcs.stat(filename)
+        except gcs.errors.NotFoundError as e:
+            stat = None
+
+        if stat is None or clobber is True:
+            gcs_file = gcs.open(filename, "w", content_type="image/png")
+            gcs_file.write(image)
+            gcs_file.close()
+            blobstore_filename = "/gs%s" % filename
+        else:
+            blobstore_filename = "/gs%s" % stat.filename
 
         # Get a URL for it from the cloud storage
-        blobstore_filename = "/gs%s" % filename
         blob_key = blobstore.create_gs_key(blobstore_filename)
         return images.get_serving_url(blob_key, size=size, secure_url=True)
